pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.121.1",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "bcrypt>=4.0.1,<4.1",
    "pydantic-settings>=2.1",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.9",
    "pyjwt[crypto]>=2.9",
    "typer>=0.12",
    "cryptography>=42.0",
    "boto3>=1.34",
    "email-validator>=2.2",
]
